using break/else defining what each get function returns
